Refactor link parsing
I've moved the parsing algorithm out of BrowseFilter to a separate class
so it was possible to introduce unit tests. There are now 68 tests which
result in 100% test coverage of the new LinkParser.

Only behavior change in the app introduced by this commit is that
"https:github.com" url now opens the main application screen instead of
a browser.